+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/PurchaseOrderInvoiceServiceImpl.java
@@ -22,13 +22,19 @@ import com.axelor.apps.account.db.InvoiceLine;
 import com.axelor.apps.account.db.repo.InvoiceRepository;
 import com.axelor.apps.account.service.invoice.InvoiceService;
 import com.axelor.apps.account.service.invoice.generator.InvoiceGenerator;
+import com.axelor.apps.account.service.invoice.generator.InvoiceLineGenerator;
 import com.axelor.apps.base.db.Product;
 import com.axelor.apps.base.service.AddressService;
 import com.axelor.apps.purchase.db.PurchaseOrder;
 import com.axelor.apps.purchase.db.PurchaseOrderLine;
+import com.axelor.apps.purchase.db.repo.PurchaseOrderRepository;
+import com.axelor.apps.supplychain.db.Timetable;
+import com.axelor.apps.supplychain.db.repo.TimetableRepository;
 import com.axelor.apps.supplychain.exception.IExceptionMessage;
+import com.axelor.apps.supplychain.service.app.AppSupplychainService;
 import com.axelor.apps.supplychain.service.invoice.generator.InvoiceGeneratorSupplyChain;
 import com.axelor.apps.supplychain.service.invoice.generator.InvoiceLineGeneratorSupplyChain;
+import com.axelor.common.ObjectUtils;
 import com.axelor.db.JPA;
 import com.axelor.exception.AxelorException;
 import com.axelor.exception.db.repo.TraceBackRepository;
@@ -40,7 +46,9 @@ import java.lang.invoke.MethodHandles;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import javax.persistence.Query;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -52,6 +60,10 @@ public class PurchaseOrderInvoiceServiceImpl implements PurchaseOrderInvoiceServ
   @Inject private InvoiceService invoiceService;
 
   @Inject private InvoiceRepository invoiceRepo;
+
+  @Inject protected TimetableRepository timetableRepo;
+
+  @Inject protected AppSupplychainService appSupplychainService;
 
   @Override
   @Transactional(rollbackOn = {Exception.class})
@@ -261,5 +273,160 @@ public class PurchaseOrderInvoiceServiceImpl implements PurchaseOrderInvoiceServ
     } else {
       return BigDecimal.ZERO;
     }
+  }
+
+  @Transactional(rollbackOn = {Exception.class})
+  protected Invoice generateInvoiceFromTimetableForPurchaseOrder(
+      PurchaseOrder purchaseOrder, List<Long> timetableIdList) throws AxelorException {
+    if (ObjectUtils.isEmpty(timetableIdList)) {
+      throw new AxelorException(
+          purchaseOrder,
+          TraceBackRepository.CATEGORY_INCONSISTENCY,
+          I18n.get(IExceptionMessage.SO_INVOICE_NO_TIMETABLES_SELECTED));
+    }
+    BigDecimal percentSum = BigDecimal.ZERO;
+    List<Timetable> timetableList = new ArrayList<>();
+    for (Long timetableId : timetableIdList) {
+      Timetable timetable = timetableRepo.find(timetableId);
+      timetableList.add(timetable);
+      percentSum = percentSum.add(timetable.getPercentage());
+    }
+    Invoice invoice = generateInvoiceFromLines(purchaseOrder, percentSum);
+
+    for (Timetable timetable : timetableList) {
+      timetable.setInvoice(invoice);
+      timetable.setInvoiced(true);
+      timetableRepo.save(timetable);
+    }
+    return invoice;
+  }
+
+  public Invoice generateInvoiceFromLines(PurchaseOrder purchaseOrder, BigDecimal percentSum)
+      throws AxelorException {
+
+    if (percentSum.equals(BigDecimal.ZERO)) {
+      throw new AxelorException(
+          purchaseOrder,
+          TraceBackRepository.CATEGORY_INCONSISTENCY,
+          I18n.get(IExceptionMessage.SO_INVOICE_NO_LINES_SELECTED));
+    }
+
+    Map<Long, BigDecimal> qtyToInvoiceMap = new HashMap<>();
+
+    for (PurchaseOrderLine purchaseOrderLine : purchaseOrder.getPurchaseOrderLineList()) {
+      BigDecimal realQty =
+          purchaseOrderLine
+              .getQty()
+              .multiply(percentSum)
+              .divide(new BigDecimal("100"), 2, RoundingMode.HALF_EVEN);
+      qtyToInvoiceMap.put(purchaseOrderLine.getId(), realQty);
+
+      if (qtyToInvoiceMap.get(purchaseOrderLine.getId()).compareTo(purchaseOrderLine.getQty())
+          > 0) {
+        throw new AxelorException(
+            purchaseOrderLine,
+            TraceBackRepository.CATEGORY_INCONSISTENCY,
+            I18n.get(IExceptionMessage.PO_INVOICE_QTY_MAX));
+      }
+    }
+    return this.generateInvoice(
+        purchaseOrder, purchaseOrder.getPurchaseOrderLineList(), qtyToInvoiceMap);
+  }
+
+  @Transactional(rollbackOn = {Exception.class})
+  public Invoice generateInvoice(
+      PurchaseOrder purchaseOrder,
+      List<PurchaseOrderLine> purchaseOrderLinesSelected,
+      Map<Long, BigDecimal> qtyToInvoiceMap)
+      throws AxelorException {
+
+    Invoice invoice =
+        this.createInvoice(purchaseOrder, purchaseOrderLinesSelected, qtyToInvoiceMap);
+    invoiceRepo.save(invoice);
+
+    Beans.get(PurchaseOrderRepository.class).save(fillPurchaseOrder(purchaseOrder, invoice));
+
+    return invoice;
+  }
+
+  public Invoice createInvoice(
+      PurchaseOrder purchaseOrder,
+      List<PurchaseOrderLine> purchaseOrderLineList,
+      Map<Long, BigDecimal> qtyToInvoiceMap)
+      throws AxelorException {
+
+    InvoiceGenerator invoiceGenerator = this.createInvoiceGenerator(purchaseOrder);
+
+    Invoice invoice = invoiceGenerator.generate();
+
+    invoiceGenerator.populate(
+        invoice, this.createInvoiceLines(invoice, purchaseOrderLineList, qtyToInvoiceMap));
+
+    return invoice;
+  }
+
+  public List<InvoiceLine> createInvoiceLines(
+      Invoice invoice,
+      List<PurchaseOrderLine> purchaseOrderLineList,
+      Map<Long, BigDecimal> qtyToInvoiceMap)
+      throws AxelorException {
+
+    List<InvoiceLine> invoiceLineList = new ArrayList<>();
+    for (PurchaseOrderLine purchaseOrderLine : purchaseOrderLineList) {
+
+      if (qtyToInvoiceMap.containsKey(purchaseOrderLine.getId())) {
+        List<InvoiceLine> invoiceLines =
+            this.createInvoiceLine(
+                invoice, purchaseOrderLine, qtyToInvoiceMap.get(purchaseOrderLine.getId()));
+        invoiceLineList.addAll(invoiceLines);
+        purchaseOrderLine.setInvoiced(true);
+      }
+    }
+
+    return invoiceLineList;
+  }
+
+  public List<InvoiceLine> createInvoiceLine(
+      Invoice invoice, PurchaseOrderLine purchaseOrderLine, BigDecimal qtyToInvoice)
+      throws AxelorException {
+
+    Product product = purchaseOrderLine.getProduct();
+
+    InvoiceLineGenerator invoiceLineGenerator =
+        new InvoiceLineGeneratorSupplyChain(
+            invoice,
+            product,
+            purchaseOrderLine.getProductName(),
+            purchaseOrderLine.getDescription(),
+            qtyToInvoice,
+            purchaseOrderLine.getUnit(),
+            purchaseOrderLine.getSequence(),
+            false,
+            null,
+            purchaseOrderLine,
+            null,
+            false,
+            null) {
+
+          @Override
+          public List<InvoiceLine> creates() throws AxelorException {
+
+            InvoiceLine invoiceLine = this.createInvoiceLine();
+
+            List<InvoiceLine> invoiceLines = new ArrayList<>();
+            invoiceLines.add(invoiceLine);
+
+            return invoiceLines;
+          }
+        };
+
+    return invoiceLineGenerator.creates();
+  }
+
+  public PurchaseOrder fillPurchaseOrder(PurchaseOrder purchaseOrder, Invoice invoice) {
+
+    purchaseOrder.setOrderDate(appSupplychainService.getTodayDate());
+
+    return purchaseOrder;
   }
 }