+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Unreleased 5.2.13]
 ## Improvements
 - OPPORTUNITY: Change the data type of bestCase and worstCase fields from string to decimal.
+- INVOICE : generate invoice from timetableList from purchase order.
 
 ## Bug Fixes
 - MRP : Remove duplicate information